+++ src/PartnerCenter/Invoices/IEstimateLinkCollectionByCurrency.cs
@@ -1,5 +1,5 @@
 ﻿// -----------------------------------------------------------------------
-// <copyright file="IEstimateLink.cs" company="Microsoft">
+// <copyright file="IEstimateLinkCollectionByCurrency.cs" company="Microsoft">
 //     Copyright (c) Microsoft Corporation. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------