+++ src/PartnerCenter/Invoices/IReceiptCollection.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Store.PartnerCenter.Invoices
     /// <summary>
     /// Represents the operations that can be done on receipt collection.
     /// </summary>
-    public interface IReceiptCollection : IPartnerComponent<string>, IEntitySelector<IReceipt>
+    public interface IReceiptCollection : IPartnerComponent<string>, IEntitySelector<string, IReceipt>
     {
     }
 }